requirements.txt
click==6.7
CouchDB==1.1
python-dotenv==0.7.1
